Abstract out the ReactQuill
