+++ src/frontend/components/componentsExport.js
@@ -8,3 +8,4 @@ export { Pallete } from "./Pallete/Pallete";
 export { Note } from "./Note/Note";
 export { MasonryLayout } from "./MasonryLayout/MasonryLayout";
 export { Modal } from "./Modal/Modal";
+export { ReactQuillEditor } from "./ReactQuillEditor/ReactQuillEditor";
